Create code block if unable to find one
MdEditCodeBlock will create a code block if unable to find one to edit.
The definition of unable to find one aka if a typed code block is not
found and the range is a single empty line. The file type is asked, by
default is markdown, if it's markdown then the code is directly inserted
otherwise a fenced code block is added around the code